==> Norway_Morsa/MetData_Processing/notebooks/05_download_era5.ipynb <==
import cdsapi
import os
import shutil
import xarray as xr
import pandas as pd

# Downloading and processing of ERA5 data

This notebook downloads **hourly** ERA5 data (for all days and months) for the specified variables, geographic area and years. Results are combined into a single netCDF for further processing. Download volumes are restricted by the API and so, to avoid errors, the code here downloads one year of data at a time and then stitches the results together using xarray. This is slow, but usually reliable. The API calls used in the code below are based on the examples originally provided by Daniel [here](https://82.223.43.150/watexr/pl/back3zg8z788ddspzjaxnnhprw).

**Note:** You must first follow the instructions [here](https://cds.climate.copernicus.eu/api-how-to#install-the-cds-api-key) to create a Copernicus user account and install the `cdsapi` client. Then fill-in the details in section 1, below, and run the notebook.

For the "common papers" component of WATExR, I believe the variables currently of interest are listed in `05_era5_s5_vars_of_interest.xlsx` (at least, for the Norwegian case study).

## 1. User input

# Area of interest. Format:
# lat_min/lon_min/lat_max/lon_max
area = '59.86/10.68/59.34/11.13'

# Variables of interest. See here for details:
# https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-single-levels?tab=form
variables = ['10m_u_component_of_wind', 
             '10m_v_component_of_wind', 
             '2m_dewpoint_temperature',
             '2m_temperature', 
             'surface_pressure', 
             'surface_solar_radiation_downwards',
             'surface_thermal_radiation_downwards', 
             'total_cloud_cover', 
             'total_precipitation',
            ]

# Years of interest
years = range(1980, 2020)

# Folder in which to save annual netCDFs
raw_fold = '/home/jovyan/shared/WATExR/ERA5/raw_netcdf'

# Path for final/tidied netCDF
merged_nc = '/home/jovyan/shared/WATExR/ERA5/morsa_era5_merged.nc'

## 2. Download ERA5

**Note:** The cell below may take a *long* time to run - **expect > 24 hours** for the full dataset.

%%time

c = cdsapi.Client()

# Loop over years
for year in years:
    # Build request
    request = {
        'product_type': 'reanalysis',
        'format': 'netcdf',
        'area': area,
        'variable': variables,
        'year': year,
        'month': [
            '01', '02', '03',
            '04', '05', '06',
            '07', '08', '09',
            '10', '11', '12',
        ],
        'day': [
            '01', '02', '03',
            '04', '05', '06',
            '07', '08', '09',
            '10', '11', '12',
            '13', '14', '15',
            '16', '17', '18',
            '19', '20', '21',
            '22', '23', '24',
            '25', '26', '27',
            '28', '29', '30',
            '31',
        ],
        'time': [
            '00:00', '01:00', '02:00',
            '03:00', '04:00', '05:00',
            '06:00', '07:00', '08:00',
            '09:00', '10:00', '11:00',
            '12:00', '13:00', '14:00',
            '15:00', '16:00', '17:00',
            '18:00', '19:00', '20:00',
            '21:00', '22:00', '23:00',
        ],
    }
    
    # Get nc file
    out_path = os.path.join(raw_fold, f'era5_{year}.nc')
    c.retrieve('reanalysis-era5-single-levels',
               request,
               out_path,
              )

## 2. Merge and tidy annual netCDFs

The code below merges the annual netCDFs into a single file and changes the ERA5 varaible names to match those in the S5 dataset from UNICAN. It also resolves some inconsistencies with variable names in ERA5: netCDFs for recent years (i.e. 2019) use a different variable naming convention compared to earlier data (see [here](https://confluence.ecmwf.int/display/CKB/ERA5%3A+data+documentation#ERA5:datadocumentation-Dataupdatefrequency) for details). 

# Read var name lookup
df = pd.read_excel('05_era5_s5_vars_of_interest.xlsx')
df

# Read all .nc files
nc_paths = os.path.join(raw_fold, '*.nc')
ds = xr.open_mfdataset(nc_paths, combine='by_coords')

# Tidy variables
for var_name in ds.variables:
    if var_name[-4:] == '0005':
        # Not needed
        del ds[var_name]
    elif var_name[-4:] == '0001':
        # Recent data (2019 only). Merge with other series
        var, idx = var_name.split('_')
        ds[var] = ds[var].combine_first(ds[var_name])
        del ds[var_name]

# Rename vars to match S5
names_dict = dict(zip(df['era5_code'], df['c4r_s5_code']))
ds = ds.rename(names_dict)

# Save
ds.to_netcdf(merged_nc)

ds
